requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-admin-rangefilter==0.5.4
django-adminactions==1.8.1
idna==2.9
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
